@@ -1,20 +1,12 @@
-// See https://kit.svelte.dev/docs/types#app
+// See https://svelte.dev/docs/kit/types#app.d.ts
 // for information about these interfaces
 declare global {
 	namespace App {
 		// interface Error {}
 		// interface Locals {}
-		interface PageData {
-			flash?: { type: 'success' | 'error'; message: string };
-		}
+		// interface PageData {}
 		// interface PageState {}
 		// interface Platform {}
-		namespace Superforms {
-			type Message = {
-				type: 'error' | 'success';
-				text: string;
-			};
-		}
 	}
 }
 
