@@ -1,12 +1,2 @@
-<script lang="ts">
-	import type { PageData } from './$types';
-	import Seo from '$lib/SEO.svelte';
-	import Blocks from '$lib/directus/components/Blocks.svelte';
-
-	export let data: PageData;
-	const { site_settings, page } = data;
-	const { seo, blocks } = page;
-</script>
-
-<Seo {seo} {site_settings} />
-<Blocks data={blocks} />
+<h1>Welcome to SvelteKit</h1>
+<p>Visit <a href="https://svelte.dev/docs/kit">svelte.dev/docs/kit</a> to read the documentation</p>
